requirements.txt
astroid==3.0.1
blinker==1.6.3
click==8.1.7
dill==0.3.7
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pylint==3.0.1
python-decouple==3.8
python-dotenv==1.0.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
